refactor: extract CountingInputStream and file naming logic into dedicated services for reuse and clarity
